remove url modal
