+++ social-network-ui/src/AppRoutes.js
@@ -3,8 +3,6 @@ import { Routes, Route } from "react-router-dom";
 import HomePage from "./pages/HomePage/HomePage";
 import AuthPage from "./pages/AuthPage/AuthPage";
 import LoginPage from "./pages/LoginPage/LoginPage";
-import ResetPasswordConfirmPage from "./pages/ResetPasswordConfirmPage/ResetPasswordConfirmPage";
-import ResetPasswordPage from "./pages/ResetPasswordPage/ResetPasswordPage";
 import ProfilePage from "./pages/ProfilePage/ProfilePage";
 
 const AppRoutes = () => {
@@ -17,11 +15,6 @@ const AppRoutes = () => {
 
       {/* Auth Routes */}
       <Route path="/signIn" element={<LoginPage />} />
-      <Route path="/resetPassword" element={<ResetPasswordPage />} />
-      <Route
-        path="/resetPassword/confirm"
-        element={<ResetPasswordConfirmPage />}
-      />
 
       {/* Profile Routes */}
       <Route path="/profile/:userId" element={<ProfilePage />} />
