--- social-network-ui/src/pages/ResetPasswordConfirmPage/ResetPasswordConfirmPage.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import ModalConfirm from '../../components/ModalConfirm/ModalConfirm'
-
-
-function ResetPasswordConfirmPage() {
-
-  return (
-    <ModalConfirm />
-  )
-}
-
-export default ResetPasswordConfirmPage
