--- social-network-ui/src/pages/ResetPasswordPage/ResetPasswordPage.jsx
@@ -1,15 +1,0 @@
-import React from 'react'
-import ModalResetPassword from '../../components/ModalResetPassword/ModalResetPassword';
-
-
-
-
-function ResetPasswordPage() {
-  
-  return (
-    <ModalResetPassword />
-
-  )
-}
-
-export default ResetPasswordPage
